Added splitting segments.
This includes some other modifications: Paths were updated to use the built-in
Segment linked list structure rather than an array, with corresponding
modifications in Image and Viewport which references them. The context menu was
updated with this and "Flip Arc", which I forgot to test, but should allow using
the other arc of an ellipse.